# vocabularySwagger.yaml
swagger: "2.0"
info:
  description: "Esta es una descripción de las URL del micro servicio Vocabulario del proyecto dentro de la materia Arquitectura de software, desarrollado por los estudiantes: miembros del grupo 1B"
  version: "1.0.0"
  title: "Proyecto Grupo 1B"
  contact:
    email: "mcarbonells@unal.edu.co"
host: "localhost"
...
schemes:
- "http"
paths:
  /category/newCategory:
    post:
      tags:
      - "category"
      summary: "Añadir una nueva categoria"
      description: ""
      operationId: "addCategory"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "categoria a añadir en la DB pidiendo name y level"
        required: true
        schema:
          $ref: "#/definitions/category"
      responses:
        "201":
          description: "successful operation"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/category"
...
  /category/updateCategory:
    put:
      tags:
      - "category"
      summary: "Modificar una categoria ya guardada"
      description: ""
      operationId: "updateC"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "categoria que se actualiza en la DB con name"
        required: true
        schema:
          $ref: "#/definitions/category"
      responses:
        "201":
          description: "successful operation"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/category"
        "404":
          description: "not found"
  /category/deleteCategory: 
    delete:
      tags:
      - "category"
      summary: "Eliminar categoria"
      description: ""
      operationId: "deleteCategory"
      parameters:
        - in: "body"
          name: "body"
          description: "Categoria que se elimina en la DB con name"
          required: true
          schema:
            $ref: "#/definitions/category"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "object"
            items:
              $ref: "#/definitions/category"
        "404":
          description: "not found"
...
definitions:
  category:
    type: "object"
    required:
    - "name"
    - "level"
    properties:
      name:
        type: "string"
        example: "Animals"
      level:
        type: "string"
        example: "A1"
    xml:
      name: "category"
...